refactor: ContextMenu
